pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yc-applier"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.44.0",
    "anthropic>=0.28.0",
    "openai>=1.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pypdf>=4.2.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
yc-apply = "yc_applier.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["yc_applier*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
